+++ packages/ui-plugins/live-activities/src/components/live-activity/common/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Common components for Live Activity features
+ * Shared between launch-live-activity and update-activity components
+ */
+
+export { ErrorMessage } from './ErrorMessage';
+export { JsonEditor } from './JsonEditor';
+export { DialogActions } from './DialogActions';
+
